% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plate_input.R
\name{check_plate_input}
\alias{check_plate_input}
\title{checks plate input for dodgy well plate combinations}
\usage{
check_plate_input(well, plate)
}
\arguments{
\item{well}{vector of well labels}

\item{plate}{integer, number of wells in full plate}
}
\description{
checks plate input for dodgy well plate combinations
}
