% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_map.R
\name{is_1536}
\alias{is_1536}
\title{internal 1536 plate function for plate_map}
\usage{
is_1536(well)
}
\arguments{
\item{well}{vector of alphanumeric well labels}
}
\description{
internal 1536 plate function for plate_map
}

