% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrapper.R
\name{is_old_ggplot}
\alias{is_old_ggplot}
\title{check ggplot2 version}
\usage{
is_old_ggplot()
}
\description{
after ggplot2 v3.3.0, using scale_y_reverse() also reverses the order
of the ylim arguments in coord_fixed()
}
