% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrapper.R
\name{legend_title}
\alias{legend_title}
\title{change legend title}
\usage{
legend_title(title)
}
\arguments{
\item{title, }{string new title}
}
\value{
ggplot object
}
\description{
Change the legend title. This can be done in ggplot but there
are a million incomprehensible ways to do it.
}
