% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_smooth.R
\name{med_smooth}
\alias{med_smooth}
\title{2-way median smooth}
\usage{
med_smooth(
  platemap,
  plate,
  eps = 0.01,
  maxiter = 10,
  trace.iter = FALSE,
  na.rm = TRUE,
  normalise = FALSE
)
}
\arguments{
\item{platemap}{dataframe produced by \code{plate_map}}

\item{plate}{numeric, number of wells in plate, either 96 or 384}

\item{eps}{real number greater than 0. A tolerance for divergence}

\item{maxiter}{int, the maximum number of iterations}

\item{trace.iter}{Boolean, should progress in convergence be reported?}

\item{na.rm}{Boolean, should missing values be removed?}

\item{normalise}{Boolean, should the data be divided by the MAD?}
}
\value{
A dataframe consisting of two column, wellID and
        polished numeric values
}
\description{
Given a platemap produced by \code{plate_map}, will return
a dataframe with after values have been transformed into
a matrix mirroring the plate structure and undergoing a
2-way median polish to remove row or column effects
}
