% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_smooth.R
\name{plate_effect}
\alias{plate_effect}
\title{Two way-median smooth on a plate map}
\usage{
plate_effect(platemap, plate)
}
\arguments{
\item{platemap}{platemap produced by \code{plate_map}}

\item{plate, }{integer either 96 or 384}
}
\description{
Given a platemap produced by \code{plate_map}, this will perform
a two way median smooth, and return the results of \code{medpolish}.
Useful for row and column effects, as well as the raw residuals.
}

