% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_map.R
\name{plate_map}
\alias{plate_map}
\title{creates dataframe of row,column,data from wellID and data}
\usage{
plate_map(data, well)
}
\arguments{
\item{data}{numeric data to be used as colour scale}

\item{well}{alpha-numeric well IDs, e.g 'A01'}
}
\value{
dataframe
}
\description{
internal function
}

