% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_map.R
\name{plate_map_grid}
\alias{plate_map_grid}
\title{creates dataframe of row, column, plate_id from data regarding wellIDs}
\usage{
plate_map_grid(data, well, plate_id)
}
\arguments{
\item{data}{numerical data to be used as colour scale}

\item{well}{alpha-numeric wellIDs, e.g 'A01'}

\item{plate_id}{plate identifers e.g 'plate_1'}
}
\value{
dataframe
}
\description{
internal function
}

