% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_map.R
\name{plate_map_multiple}
\alias{plate_map_multiple}
\title{row, column for multiple features}
\usage{
plate_map_multiple(data, well)
}
\arguments{
\item{data}{vector or dataframe of numeric data}

\item{well}{vector of alphanumeric well IDs e.g 'A01'}
}
\description{
Generates a dataframe for multiple features, given a wellID column and multiple
features
}

