% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_map.R
\name{plate_map_scale}
\alias{plate_map_scale}
\title{creates dataframe of row, column, and scaled data from well IDs}
\usage{
plate_map_scale(data, well)
}
\arguments{
\item{data}{numeric data to be used as colour scale}

\item{well}{alpha-numeric well IDs, e.g 'A01'}
}
\value{
dataframe
}
\description{
internal function
}

