% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrapper.R
\name{plt1536}
\alias{plt1536}
\title{ggplot plate object}
\usage{
plt1536(platemap, size = 3.5, shape = 22)
}
\arguments{
\item{platemap}{platemap dataframe produced by \code{plate_map}}

\item{size}{int, size parameter for ggplot2::geom_point}

\item{shape}{int, shape parameter for ggplot2::geom_point}
}
\value{
ggplot object
}
\description{
internal function
}
