% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrapper.R
\name{plt384}
\alias{plt384}
\title{ggplot plate object}
\usage{
plt384(platemap)
}
\arguments{
\item{platemap}{platemap dataframe produced by \code{plate_map}}
}
\value{
ggplot object
}
\description{
internal function
}

