% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrapper.R
\name{plt6}
\alias{plt6}
\title{ggplot plate object}
\usage{
plt6(
  platemap,
  size = 50,
  shape = 21,
  na_fill = "white",
  na_alpha = 0.1,
  na_size_ratio = 0.9
)
}
\arguments{
\item{platemap}{platemap dataframe produced by \code{plate_map}}

\item{size}{int, size parameter for ggplot2::geom_point}

\item{shape}{int, shape parameter for ggplot2::geom_point}

\item{na_fill}{string, fill colour for na or missing values}

\item{na_alpha}{float, alpha transparancy for missing or na values}

\item{na_size_ratio}{float, size ratio for missing values, set to 1 for same
size as normal values.}
}
\value{
ggplot object
}
\description{
internal function
}
