% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_grid.R
\name{raw_grid}
\alias{raw_grid}
\title{Plots multiple platemaps with heatmap of raw values}
\usage{
raw_grid(data, well, plate_id, ncols = 2, plate = 96)
}
\arguments{
\item{data}{Numerical values to be plotted}

\item{well}{Vector of well identifiers e.g "A01"}

\item{plate_id}{Vector of plate identifiers e.g "Plate_1"}

\item{ncols}{Number of columns to display multiple heatmaps}

\item{plate}{Number of wells in complete plate (96, 384 or 1536)}
}
\value{
ggplot plot
}
\description{
Converts numerical values. well labels, and plate labels into multiple 
plate heatmaps
}
\examples{
df01 <- data.frame(well = num_to_well(1:96),
  vals = rnorm(96),
  plate = 1)

df02 <- data.frame(well = num_to_well(1:96),
  vals = rnorm(96),
  plate = 2)

df <- rbind(df01, df02)

raw_grid(data = df$vals,
    well = df$well,
    plate_id = df$plate,
    plate = 96)
}

