% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{readmap_data}
\alias{readmap_data}
\title{example data in a plate map form}
\format{96 integers structured in a the form of a 96-well plate}
\source{
none
}
\usage{
readmap_data
}
\description{
example data in a plate map form
}
\keyword{datasets}

