% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_plate.R
\name{rotate_plate}
\alias{rotate_plate}
\title{rotates matrix by 180 degrees}
\usage{
rotate_plate(m)
}
\arguments{
\item{m}{matrix}
}
\value{
matrix
}
\description{
If someone (no names) puts in a plate upside down, this function 
will rotate a plate matrix produced by \code{plate_matrix} to be
the correct way up. I.e if A01 is in the bottom right hand corner rather
than the top left.
}

