% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrapper.R
\name{plt1536}
\alias{plt1536}
\title{ggplot plate object}
\usage{
plt1536(platemap)
}
\arguments{
\item{platemap}{platemap dataframe produced by \code{plate_map}}
}
\value{
ggplot object
}
\description{
internal function
}

